+++ utility/gov.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"math/big"
 
+	coreTypes "github.com/pokt-network/pocket/shared/core/types"
 	typesUtil "github.com/pokt-network/pocket/utility/types"
 	"google.golang.org/protobuf/types/known/wrapperspb"
 )
@@ -494,7 +495,7 @@ func (u *UtilityContext) GetParamOwner(paramName string) ([]byte, error) {
 	}
 }
 
-func (u *UtilityContext) GetFee(msg typesUtil.Message, actorType typesUtil.ActorType) (amount *big.Int, err typesUtil.Error) {
+func (u *UtilityContext) GetFee(msg typesUtil.Message, actorType coreTypes.ActorType) (amount *big.Int, err typesUtil.Error) {
 	switch x := msg.(type) {
 	case *typesUtil.MessageDoubleSign:
 		return u.GetMessageDoubleSignFee()
@@ -502,52 +503,52 @@ func (u *UtilityContext) GetFee(msg typesUtil.Message, actorType typesUtil.Actor
 		return u.GetMessageSendFee()
 	case *typesUtil.MessageStake:
 		switch actorType {
-		case typesUtil.ActorType_App:
+		case coreTypes.ActorType_ACTOR_TYPE_APP:
 			return u.GetMessageStakeAppFee()
-		case typesUtil.ActorType_Fisherman:
+		case coreTypes.ActorType_ACTOR_TYPE_FISH:
 			return u.GetMessageStakeFishermanFee()
-		case typesUtil.ActorType_ServiceNode:
+		case coreTypes.ActorType_ACTOR_TYPE_SERVICENODE:
 			return u.GetMessageStakeServiceNodeFee()
-		case typesUtil.ActorType_Validator:
+		case coreTypes.ActorType_ACTOR_TYPE_VAL:
 			return u.GetMessageStakeValidatorFee()
 		default:
 			return nil, typesUtil.ErrUnknownActorType(actorType.String())
 		}
 	case *typesUtil.MessageEditStake:
 		switch actorType {
-		case typesUtil.ActorType_App:
+		case coreTypes.ActorType_ACTOR_TYPE_APP:
 			return u.GetMessageEditStakeAppFee()
-		case typesUtil.ActorType_Fisherman:
+		case coreTypes.ActorType_ACTOR_TYPE_FISH:
 			return u.GetMessageEditStakeFishermanFee()
-		case typesUtil.ActorType_ServiceNode:
+		case coreTypes.ActorType_ACTOR_TYPE_SERVICENODE:
 			return u.GetMessageEditStakeServiceNodeFee()
-		case typesUtil.ActorType_Validator:
+		case coreTypes.ActorType_ACTOR_TYPE_VAL:
 			return u.GetMessageEditStakeValidatorFee()
 		default:
 			return nil, typesUtil.ErrUnknownActorType(actorType.String())
 		}
 	case *typesUtil.MessageUnstake:
 		switch actorType {
-		case typesUtil.ActorType_App:
+		case coreTypes.ActorType_ACTOR_TYPE_APP:
 			return u.GetMessageUnstakeAppFee()
-		case typesUtil.ActorType_Fisherman:
+		case coreTypes.ActorType_ACTOR_TYPE_FISH:
 			return u.GetMessageUnstakeFishermanFee()
-		case typesUtil.ActorType_ServiceNode:
+		case coreTypes.ActorType_ACTOR_TYPE_SERVICENODE:
 			return u.GetMessageUnstakeServiceNodeFee()
-		case typesUtil.ActorType_Validator:
+		case coreTypes.ActorType_ACTOR_TYPE_VAL:
 			return u.GetMessageUnstakeValidatorFee()
 		default:
 			return nil, typesUtil.ErrUnknownActorType(actorType.String())
 		}
 	case *typesUtil.MessageUnpause:
 		switch actorType {
-		case typesUtil.ActorType_App:
+		case coreTypes.ActorType_ACTOR_TYPE_APP:
 			return u.GetMessageUnpauseAppFee()
-		case typesUtil.ActorType_Fisherman:
+		case coreTypes.ActorType_ACTOR_TYPE_FISH:
 			return u.GetMessageUnpauseFishermanFee()
-		case typesUtil.ActorType_ServiceNode:
+		case coreTypes.ActorType_ACTOR_TYPE_SERVICENODE:
 			return u.GetMessageUnpauseServiceNodeFee()
-		case typesUtil.ActorType_Validator:
+		case coreTypes.ActorType_ACTOR_TYPE_VAL:
 			return u.GetMessageUnpauseValidatorFee()
 		default:
 			return nil, typesUtil.ErrUnknownActorType(actorType.String())